Fix the Travis CI-CD errors
**Why was this change made?**
Travis CI example app build job is failing because of the following
problems:
doc-lib module node dependencies: They had to be manually installed by
going into the doc-lib project folder creating extra complication to the
script phase of doc-lib package build. So those dependencies are
duplicated in the main package.json dependencies so that they can be
installed as main node modules and any executables can be executed right
from the main node-module's .bin

Compodoc documentation generation: The components build scripts are
modified to use the compodoc executable in the main node_module's json

AOT build of examples: This is failing right now with a strange error:
Please add doc container component to a module. So using regular build
for the examples.

**Testing Done:**
- Ran a Travis CI pipeline on forked repo and made sure it passed
- Started the examples website on local server and saw that it is
working as expected

Signed-off-by: ps37 <prudhvi.af121@gmail.com>