Implement clipped text directive (#42)
* Implement clipped text directive

Changed from a component into a directive to avoid creating unnecessary DOM nodes
and to make its API simpler. It also makes it easier to dynamically create it since
the component version required trancluded nodes that are complicated do do when
rendering components dynamically as we do on our grid.

Remove usage of vcd-clipped-text since it's been deprecated

This does not make it work as expected. Fixing it will be taken care of by https://jira.eng.vmware.com/browse/VDUCC-54

Signed-off-by: Juan Mendes <jmendes@vmware.com>