Substitute `;` with `&&` when running npm scripts
This change make the script fails at the first error met.
When `;` is used and a command fails the execution will continue
with the next one and so on, thus the last visible output will
contain irrelevant error and not the root one. `&&` causes the script
to stop the moment an errot occurs

Signed-off-by: Ivo Rahov <irahov@vmware.com>