Fix SpotlightSearchComponent to be able to auto-select item from a second section
Problem description:
Consider to synchronous search providers. When there is a search and both
return values, then the first match from the first section is auto-selected.
But when there is a subsequent search that yields matches only from the
second provider then no preselection occurs.

The reason is that during the second run:
1. this selectedItem is nullified
2. go through the search providers getting search matches
3. after the search from the first provider is done it does not return a result,
  But the selectFirst function was written in a way that it traverses the next
  synchronous provider and its first item is taken for preselection, although this item is
  from the first search run
4. The program goes on and after the search from the second provider is taken, the selectFirst
  is called again, but it returns immediately since
  there is a selectedItem and it is not aware that this selctedItem is no more
  valid (since it is from the first search run)

The solution is to mark all the searchSections that isLoading is in progress.
In that way the selectFirst may behave correctly for the second provider

Signed-off-by: Ivo Rahov <irahov@vmware.com>