Enable projecting content when there is no search
There is quite much empty space when there is no search which
to be utilized in one way or another.

One example is to display some helpful information on how the spotlight search can be used.

Testing done:
By using the spotlight search examples add text which is shown when there is
search, i.e. the search input is empty.

Signed-off-by: Ivo Rahov <irahov@vmware.com>