Display total number of results per search provider.
The search result of a provider can return a subsection of the total number
of items that match the criteria. Such an example is a backend quiry. The ideal
behavior is to be able to load the next page but we are still not there to implement this.

The current solution for this scenario is to show the number of the found items,
the total number of items and a warning message to the user.

The changes include:
- modify the QuickSearchResult to contain items, page, pageSize and total
- update the search component to
  -- show the partial results found
  -- display a warning message if there are more items
- add new SearchProvider for the sync-async example

Signed-off-by: Ivo Rahov <irahov@vmware.com>