Always show section name in the quick search
It is better to always show the section name even if there are no
results found. Otherwise if a section is hidden dynamically
the user may not know what exactly has happened - if there are no results or
a buggy behavior.

As a consequence when there are no matching items found
'No results found' is displayed for the corresponding section

Also the 'no title example' is removed.

Signed-off-by: Ivo Rahov <irahov@vmware.com>