Modify sendKeyboardEvent to be more versatile
Some libraries like mousetrap use deprecated, read-only keyboard event properties
like `which`. This makes it impossible to use the general `new KeyboardEvent`
mechanism of creating the event and dispatching it later. To overcome this
we have used `Object.defineProperty`.

In addition to this, the method signature was also changed to allow more versatile
approach of setting different event properties.

Signed-off-by: Ivo Rahov <irahov@vmware.com>