Fix Quick Search 'Content Projection' example not opening on Stackblitz
The Quick Search 'Content Projection' example was not opening on stackblitz with an error in the console.
In order to run an example on stackblitz the component name should be matching its module name in the
following manner:
componentName.replace('ExampleComponent', 'ExampleModule')
In this specific case the name of the component was missing the 'Example' part, i.e. it was
`QuickSearchContentProjectionComponent` and for that reason the name replace was not working
correctly hence the error, hence the failure to open the example on stackblitz

Signed-off-by: Ivo Rahov <irahov@vmware.com>