Add hideWhenEmpty flag to QuickSearchProvider to control section display when there is no data
`hideWhenEmpty` is an optional flag to control the visibility of the corresponding section.
It is needed because there might be cases when the app logic requires
hiding a given section when it is empty.

When set to `true` the section will not be displayed if it contains no data.
Defaults to `false` and in that case the section is always visible,
rendering 'No results found.'

Signed-off-by: Ivo Rahov <irahov@vmware.com>