Fix datagrid always returning new `datagridSelection`
If a getter returns `[]` it is considered as a new reference in the angular detect changes process
while `[]` in the template is not considered a new reference.
In other words the following 2 have different behaviour:
```
<component_tag
   [someInput]='some_getter_that_retyrns_empty_array'
></component_tag>
```
```
<component_tag
   [someInput]='[]'
></component_tag>
```
where the first one will result in setting `someInput` during the change detection
while the second one will not

So on one hand we do not want extra calculations that occur when there is
a new reference, but on the other hand we need a new reference when there
is a real change.

This CLN attempts to fix that with regards to `datagridSelection`

Signed-off-by: Ivo Rahov <irahov@vmware.com>