Fixed //@ ts-ignore issue with the Widget Object pattern
The Widget object finder was trying to imply the type of the widget object it was locating. This meant it had to both imply the widget type but also the type of the underlying driver. This was very difficult to do in Typescript and we had to put in a hacky solution.

This change removes this hacky solution and instead has the client of our widget supply the type of the widget that is being located. See review for examples.