Enabling widget objects to interact with its elements
This enables widget objects to abstract interactions such as a login
so that callers don't need to explicitly type into two fields
and hit the enter button. They can just call `widget.login(usr, pwd)`

This is the first step and we're going to do some refactoring.

The most important thing to understand is that there are two types
of wrappers around HTMLElement that we deal with:
  * An internal representation within the widget object and should
    remain private
  * An external representation that is exposed to callers, which
    will have all the methods of the underlying testing framework,
    e.g., DebugElement and Cypress.Chainable. Note that
    we wrap DebugElement with TestElement to make assertions simpler

Signed-off-by: Juan Mendes <jmendes@vmware.com>