Update the action availability to accept observables
**What was the change?**
Some of the actions, particularly the visibility static actions depends
on values other than selected entities, like the response from async
requests. For such cases we need to accept observables to keep their
visibility up to date.

**Why was this change made?**
There are some bugs in the VCD ui related to visibility of static
actions not getting updated because the visibility is not calculated
when one of the variables on which their visibility depends on is not
updated.

Tested the change on examples app. Yet to test on VCD UI

**Bug Number:**
VDUCC-555

**Pull request URL:**

Signed-off-by: ps37 <prudhvi.af121@gmail.com>