Responsive input directive now allocates 1/4th of the container space for input field labels on large modals
**Why was this change made?**
On large modals, most of the labels that are not very long are wrapping because of allocating the only 1/6th of the container space. In this change, the ResponsiveInputDirective is changed to allocate 1/4th of the space to avoid wrapping of labels.

**Testing Done:**
Tested the input labels on large modals in the following screens on Vcd UI: Notice that some of the labels on the left side of the following screenshots are not wrapping to a new line.

**Bug Number:**
VDUCC-642

Signed-off-by: ps37 <prudhvi.af121@gmail.com>