Optimize ActionMenuComponent for less processing on input change
Currently on any reference input change for `actions` and `selectedEntities`
the component is doing redundant computations and emits `actionsUpdate`
even though there is no actual change.

This input reference change can be as simple as just another empty array `[]`.
There are other cases when the `selectedEntities` are wrongly provided by a getter:
get selectedEntities() {
  return [this.currentSelection];
}
in this cases also there is no actual change, but angular will call the getter
from the template on every digest cycle, the getter will return a new reference
which will lead to performing all those calculations, emitting `actionsUpdate`
and all the subsequent calculations in the corresponding subscribers.

Although it is not good to use a function (the getter in this case) within a template
we are developing a library and cannot control every aspect of how
this library is going to be called that's why we'd like to make it as rigid as possible

For that purpose, this CLN is introducing a shallowEqual comparison for those
two inputs: `actions` and `selectedEntities` and this does quite much of
optimizations

Signed-off-by: Ivo Rahov <irahov@vmware.com>