Add examples to show different configs in which datagrid action menu can be displayed
**What was the change?**
Examples to show different configs in which datagrid action
menu can be displayed

**Why was this change made?**
The action display config input to datagrid is a nested object that is
used to display action menu inside a datagrid in different ways. And
there is no example that would help the library users to learn this
quickly.

# What testing was performed on this change?
Opened the examples app on local server and tested that the example is
rendered correctly

Signed-off-by: ps37 <prudhvi.af121@gmail.com>