Add unit test to ensure datagridSelection returns same reference
If there is no change in the datagrid selection the getter method `datagridSelection`
should return the same array reference. This is needed for performance optimization
so that angular will not trigger any unnecessary changeDetections and calculations
in the case when the array reference has been changed, but its contain is the same data
(for example empty arrays are never equal []!==[])

Ensuring `datagridSelection` returns same array reference when there is no real selection change
during angular detect changes cycles was done some time ago, but unit tests
were not written. This CLN just adds unit tests for that.

Signed-off-by: Ivo Rahov <irahov@vmware.com>