Removes internalFactory from WidgetObject
WidgetObject#internalFactory has caused confusion for developers. It caused a
different way to interact with components depending on whether you're inside a
WidgetObject or you're with a test (stepdef or unit test). To avoid this
confusion, we now always expose WidgetObjectElements and callers call unwrap()
on their own if they want to use the `T` functionality that is specific to
the testing framework being used.

Signed-off-by: Juan Mendes <mendesjuan@gmail.com>