Fix route analyzer v004
Previous version (@vcd/route-analyzer@0.0.4) had a mismatch between the version of yargs
in package.json and projects/route-analyzer/package.json

This caused version 0.0.4 to be bad since it uses the version
in projects/route-analyzer/package.json

This was not caught in development because by using npm link to tests changes, modules were being resolved from vcd-ui-common's dist folder instead of the target application, that meant it was reading from node_modules just above the dist folder which did have the correct version of yargs.

To avoid this problem, updated the README to explain that we should use npm install setting package.json's @vcd/route-analyzer entry to point to the dist/route-analyzer.

Signed-off-by: Juan Mendes <mendesjuan@gmail.com>