Adding a regex property to objects created by route-analyzer
This regex can be tested against existing URLs to see if
a URL matches a route.

This is going to be used in VCD UI to find the tagName for
any given URL

A warning is added for components without tag names. That is because
without a name tag, we can't automatically assert that
a page has been rendered. Users who want this feature
should always define a selector property on their components.

Testing Done
* Unit tests
* Installed on internal vcd_ui app and ran automated route tests
  * Provider: https://dashboard.cypress.io/projects/ecdw8h/runs/3981/overview
  * Tenant: https://dashboard.cypress.io/projects/ecdw8h/runs/3982/overview

Signed-off-by: Juan Mendes <mendesjuan@gmail.com>