Changing publishing to be based on magic strings in commit message
Checking the actual package.json version was causing problems
when using versions such as `3.0.0-dev.1` because the action we used
wasn't able to tell that `3.0.0-dev.2` was higher

We don't check versions anymore. Users are responsible for updating
package.json to a valid version and an error will occur if, for
example, they add the magic string but forget to bump package.json.

The magic strings are described in the Publishing section of our
README.md file

Signed-off-by: Juan Mendes <mendesjuan@gmail.com>