Truncate checkbox text if it is too long and show the whole text in tooltip
Wrap the checkbox text within a span and add vcdShowClippedText directive to it.
Change `label.clr-control-label` display type from `inline-block` to `flex` in order to make its child (the span element) block-flavoured.

Testing Done:
Built and manually tested within VCD.
Confirmed that the label is correctly truncated and a tooltip with its whole content is visible on hover when the checkbox have limited width.

Changelog is updated

Signed-off-by: Presiana Dimitrova <dpresiana@vmware.com>