Fix vcd-action-menu matching the wrong content-area
BrowserAnimationsModule causes the angular routing to delay removing old DOM elements that were present in previous routes. This leads to cases where we have two '.content-area' elements, and the DynamicDropdownPositionDirective matches the one which later gets destroyed.

This commit makes the DynamicDropdownPositionDirective compute the contentAreaElement just before positioning the dropdown menu, ensuring it picks the relevant '.content-area' at the time of its opening.

Signed-off-by: Davi Barreto <dbarreto@vmware.com>