Making components publishable
Use nx'x packager instead of ng.

Adjusted outputs in `18n` and `components` since nx wasn't
able to find the output of `i18n` though it was able to
detect that `i18n` needed to be built and built it.

Testing Done:

Run `nx build components` making sure that dist was empty,
ensuring that `i18n` and `components` were built.

Signed-off-by: Juan Mendes <mendesjuan@gmail.com>