Making publishing be manual GitHub Workflows
Now there's no need to remember magic strings in the commit message.
We can just commit the changes to package.json and CHANGELOGs and
then manually publish any package.

Since there's no way to test it until it's in main. The task is going
to be a noop using `npm publish --dry-run` so that we can test it
and fix any bugs.

See https://docs.github.com/en/actions/managing-workflow-runs/manually-running-a-workflow

Signed-off-by: Juan Mendes <mendesjuan@gmail.com>