Fix deploying app
Running `nx build examples` was no longer working because the tasks
from examples did not properly `dependsOn` "^build".
It wrongly assumed that tasks in the array `dependsOn` would be run
serially.

Had an error building because the update of package-lock.json
pulled a new version of @cds that required `@babel/runtime`

Signed-off-by: Juan Mendes <mendesjuan@gmail.com>