Associate search input with search results for accessibility purposes
The screen reader isn't able to navigate to search results in
the quick-search component and isn't reading the selected item by arrows
navigation.

The aria-activedescendant property provides an alternative method of
managing focus. It moves just a screen reader's focus to the selected
element.
The aria-activedescendant.directive generates a unique id for the
applied component and marks it as an active descendant if it's true.
The service exists to create a communication way between the directive
and an ascendant.
The results grid will be destroyed when search input is empty, because
there is a case when it's empty but the headings are actually inside a
dom, and screen reader is announcing them.

Semantic changes:
- The newly added roles such as "grid", "rowgroup", "row", "gridcell"
are making it possible to iterate over the search results with a screen
reader.
- The button element inside the .search-result-item is to mark an item
as clickable, it's not just a menu or list item.
- The empty span element inside the button element in the
.search-result-item has labeledby property which points to a parent
group and section headings. The screen reader is able to read the
headings with the selected item content with this span element.

Testing Done:
The best way to try changes is inside the live-examples > quick search
> quick search nested providers.
- Turn on your screen reader
- Open the quick search modal
- Type something inside the search input
Expected:
- Screen reader to read the preselected search result
- Screen reader to read the selected item on the arrows navigation
- User to be able to iterate over the search results with TAB button

Signed-off-by: Dmytro Bohdiev <dbohdiev@vmware.com>