Fix accessibility of pin button in quick-search.component
The 'pin' button has no label and no state, the screen reader is
announcing this element as 'button' and doesn't announce the state of
the button.

The focus on pin action is moved to pin button, because it makes it more
obvious and makes screen reader to read the button state.

Testing Done:
- Serve examples
- Open quickSearch/example/filters
- Open quick-search modal
- Turn on the screen reader
- Navigate to the pin button
Expected screen reader to announce 'Pin button'
- Click on pin button
Expected screen reader to announce button name and state

Signed-off-by: Dmytro Bohdiev <dbohdiev@vmware.com>