Add a number of results for screen readers
It is expected screen reader reads the search results amount when it
is changed or the search performed.

The change will add a new visually hidden (but not for screen readers)
live aria with aria-live=polite attribute to make screen readers
announce any change related to the results amount.
Aria-live indicates that element will be updated and the politeness of
the updates.

Testing Done:
- Open /quickSearch/example/nested
- Open the quick-search modal
- Turn on your's screen reader
- Perform the search
Expected screen reader to announce the results number.

Signed-off-by: Dmytro Bohdiev <dbohdiev@vmware.com>