[VACE-4391] Extensibility framework cause the actions menu to be redrawn numerous times causing tens/hundreds API calls to external systems
[publish @vcd/ui-components@latest]

Issue:
When the action menus in vcd get refreshed with new items (replacing)
the action menu "jumps" because there is not enough space for the
action menu to be redrawn.

Fix:
If there is not enough space for the action menu to be redrawn,
show it at the middle of the screen

Testing Done:
- Install common components to vcd and verify the action menu position
remains constant  if it doesn't have enough space.

Signed-off-by: Nikola Vladimirov Iliev <nvladimirovi@vmware.com>