Revert behavior of SelectorUtil.extractSelector
Erroneously assumed that the selector passed in would always contain
a `dataUiSelector` or a `cssSelector` but that is not the case. Instead,
just change the signature to return `undefined | string` and handle it
in the callers.

The angular implementation required an update since it would error if the same
case occurred, that is, calling WidgetObjectElement.get() passing it an object
that doesn't contain a selector.

Signed-off-by: Juan Mendes <mendesjuan@gmail.com>