Fix tooltip keyboard a11y
The element with the assigned tooltip has to be accessible through the
tab key.
There are cases when users don't use a screen reader, but use the
tab key to navigate through the UI.
So if hostElement has a tooltip, it should be focusable.

Fix Description:

The directive will be observing the subtree characterData and resizing
mutations because it could lead to overflow changing. It will save
the initial tabIndex and change it just in case the element is
non-focusable, otherwise, there is no need to change it.
The changes are not needed in case the element is not overflowing.

Testing Done:

- Open examples and go to show-clipped-text-sizing
- Try changing all tree inputs and depending on overflowing they should
and shouldn't be accessible through the Tab key.

Signed-off-by: Dmytro Bohdiev <dbohdiev@vmware.com>