Fix availability functionality for Action Menu
# Do not forget to separate subject from body with a blank line
# Why was this change made?
The availability function for actions was not being used to
hide the action. For example () => false would continue
to show the action.

# What testing was performed on this change?
# Please list steps as bullet items for manual testing.
Testing Done:
Wrote some unit tests and verified that they failed
before the fix and passed after the fix.
Used the fixed version in our application and
verified that it works in our product.

# Reference to Bugzilla or JIRA issues
Bug Number: none

Signed-off-by: Bryan Bozzi <bryanv@vmware.com>