Generate constants for segments in routes
See comments in README.md.

One of the shortcomings is that the variables created are not scoped
to the route they came from. That is because if they were, the names
would have to be extremely long. For example, for a path such as
`cloud/organizations/:orgId`, and we wanted to access the orgId segment,
we would need something like the following

```
export const  segment_cloud = "organizations";
export const  segment_cloud_organizations = "organizations";
export const  segment_cloud_organizations_orgId = "orgId";
```

That would give us a little bit more safety at the expense of creating extremely
long variables. We could have suggested a pattern where they are renamed
as they are imported but we didn't think the extra safety was worth that effort.

```
import  {segment_cloud_organizations_orgId as segment_orgId} from "@my/path"
```

Signed-off-by: Juan Mendes <mendesjuan@gmail.com>