Introduce a new ``@Input() hideLabel`
In order to comply with the accessibility requirements our forms should have either `label` or `aria-label`.
We decided that we are not going to introduce a new input like `ariaLabel`. Instead, we will assume that label is required(until we upgrade to Angular 16 and be able to set it as required) and
will have a new input `hideLabel` that will be used for hiding the label.
When `hideLabel` is set to `true`, the label won't be displayed and aria-label attribute will be set with the label's value.

For backward compatibility, label will be hidden with css when no value is provided and hideLabel is not set - I would expect this to be changed once we upgrade to Angular 16 and we set the label to be required input - then user
will be forced to set the label and use hideLabel input to hide the label.

Testing Done
============
* Have a vcd-form-input/vcd-form-select/vcd-form-checkbox with `@Input() label` set and not `@Input hideLabel` set
* the input is associated with <label> element
* Have a vcd-form-input/vcd-form-select/vcd-form-checkbox with `@Input label` set and `@Input hideLabel=true` set
* the input has an `aria-label` attribute set and not label is displayed

Signed-off-by: Tsvetan Emilov <tsvetant@vmware.com>