Fix ExpressionChangedAfterItHasBeenCheckedError
Description
===========
The ExpressionChangedAfterItHasBeenCheckedError issue revealed the
following flaw:
When the ErrorBanner is closed with clearing the input `errorMessage`
it emits the error message which in the two-way binding sets the outer
errorMessage to the newly emitted value. As a consequence this leads to
the ExpressionChangedAfterItHasBeenCheckedError but in the first place
the ErrorBanner component should not emit at all in this case.
Let's explain with an example:
Some component is using the error banner as this
```
<vcd-error-banner #banner [(errorMessage)]="outerErrorMessage"></vcd-error-banner>
```
when an error has been shown and the outer component wants to hide it by
setting `outerErrorMessage` to '' this leads to the following
* sets the ErrorBanner `errorMessage` to ''
* hides the ErrorBanner
* ErrorBanner emits `undefined`
* `outerErrorMessage` is set to `undefined` (wow, but it was just set to '')

In additon to fixing the redundant emit, ErrorBanner now uses `null`
instead of `undefined` when the error message is reset.

Fix Description
===============
When closing the ErrorBanner with no value, do not emit.
The other emit is preserved - when closing the ErrorBanner by clicking
the X button it emits `null` for errorMessage.

Testing Done
============

Unit tests
--------------

Show/Hide error
--------------
* Open Dev console
* Navigate to Error Banner example page
* Click several time the `Show/Hide Error` button
  * Verify the error is shown and hidden repeatedly
  * Verify there is no  ExpressionChangedAfterItHasBeenCheckedError
   in the console

Signed-off-by: Ivo Rahov <irahov@vmware.com>