Fix CSV sanitizing in DataExporter
Description
===============
If a value contained a character that required wrapping the
value with quotes, such as a double quote, then the csv wasn't
being sanitized for formula injection.

Fix Description
=================
We are using the suggestions from https://owasp.org/www-community/attacks/CSV_Injection

* Use a single quote instead of a tab as the sanitization character
* Also trimming content because Numbers was ignoring whitespace and still executing formulas

Removed some tests which had been testing a public method that was not used.

Signed-off-by: Juan Mendes <mendesjuan@gmail.com>